Все задания Этапа 2 выполнены (5), ошибки исправлены (10)

+++ ewm-main-svc/src/main/java/ru/practicum/ewm/compilation/service/CompilationServiceImpl.java
@@ -76,6 +76,11 @@ public class CompilationServiceImpl implements CompilationService {
 
     @Override
     public void delete(Long id) {
+
+        if (!compilationRepository.existsById(id)) {
+            throw new NotFoundException("Подборка с ID " + id + " не найдена");
+        }
+
         compilationRepository.deleteById(id);
         log.info("Подборка удалена");
     }
